requirements.txt
fastapi
uvicorn[standard]
pydantic
motor           
redis          
faststream[rabbit,cli]
python-dotenv
